+++ packages/pluggableWidgets/combobox-web/src/helpers/BaseDatasourceOptionsProvider.ts
@@ -1,27 +1,40 @@
+import { debounce } from "@mendix/widget-plugin-platform/utils/debounce";
 import { ListAttributeValue, ListValue, ObjectItem } from "mendix";
 import { FilterTypeEnum } from "../../typings/ComboboxProps";
 import { BaseOptionsProvider } from "./BaseOptionsProvider";
 import { datasourceFilter } from "./datasourceFilter";
 import { CaptionsProvider, SortOrder, Status } from "./types";
 import { DEFAULT_LIMIT_SIZE } from "./utils";
+import { FilterCondition } from "mendix/filters";
 
 export interface BaseProps {
     attributeId?: ListAttributeValue["id"];
     ds: ListValue;
     filterType: FilterTypeEnum;
     lazyLoading: boolean;
+    filterInputDebounceInterval?: number;
 }
 
 export class BaseDatasourceOptionsProvider extends BaseOptionsProvider<ObjectItem, BaseProps> {
     private ds?: ListValue;
     private attributeId?: ListAttributeValue["id"];
     protected loading: boolean = false;
+    private debouncedSetFilter: (filterCondition: FilterCondition | undefined) => void;
+    private filterInputDebounceInterval: number;
 
     constructor(
         caption: CaptionsProvider,
-        protected valuesMap: Map<string, ObjectItem>
+        protected valuesMap: Map<string, ObjectItem>,
+        filterInputDebounceInterval: number = 200
     ) {
         super(caption);
+        this.filterInputDebounceInterval = filterInputDebounceInterval;
+
+        const [debouncedFn] = debounce((filterCondition: FilterCondition | undefined) => {
+            this.ds?.setFilter(filterCondition);
+        }, this.filterInputDebounceInterval);
+
+        this.debouncedSetFilter = debouncedFn;
     }
 
     get sortOrder(): SortOrder {
@@ -51,10 +64,10 @@ export class BaseDatasourceOptionsProvider extends BaseOptionsProvider<ObjectIte
     getAll(): string[] {
         if (this.lazyLoading && this.attributeId) {
             if (this.searchTerm === "") {
-                this.ds?.setFilter(undefined);
+                this.debouncedSetFilter(undefined);
             } else {
                 const filterCondition = datasourceFilter(this.filterType, this.searchTerm, this.attributeId);
-                this.ds?.setFilter(filterCondition);
+                this.debouncedSetFilter(filterCondition);
             }
 
             return this.options;